Task 1: Digital Signature
This is the first task of the secure messaging prototype project. In this task, 4 functions were written within a wrapper class to be called in later tasks. These functions were to generate keys, hash a message, sign the hash of the message, and verify a digital signature. The python library Cryptography was used to create these functions.